#573 Add TabsTemplateEditor component and clean architecture
