+++ twake/frontend/src/app/scenes/Client/MainView/Tabs/DeprecatedTabs.js
